Updating client with getMessage method, and fixing websocket auth
